fixed rendering issues with editor community selector
